Day 17: Converted Ints to Longs

+++ aoc17/src/main/kotlin/de/havox_design/aoc2022/day17/Rock.kt
@@ -1,6 +1,6 @@
 package de.havox_design.aoc2022.day17
 
-enum class Rock(val dimensionX: Int, val dimensionY: Int, private val structure: Map<Position, String>) {
+enum class Rock(val dimensionX: Long, val dimensionY: Long, private val structure: Map<Position, String>) {
     HORIZONTAL_LINE(
         4,
         1,
